requirements.txt
tensorflow
streamlit
numpy
pillow
pandas
matplotlib
scikit-image
opencv-python
